Add maven dependencies

+++ src/main/java/app/computer_school/models/Genre.java
@@ -1,6 +1,6 @@
 package app.computer_school.models;
 
 public class Genre {
-    protected int id;
+    protected Long id;
     protected String title;
 }
